services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dance-api"
version = "0.1.0"
description = "Just-Dance MVP API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-multipart>=0.0.9",
  "numpy>=1.26.0",
  "scipy>=1.13.0",
  "librosa>=0.10.2.post1",
  "pillow>=10.3.0",
  "requests>=2.32.0",
  "google-auth>=2.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
